Sauce_demo/requirement.txt
# Core Robot Framework
robotframework==7.4.2

# Selenium library for web automation
robotframework-seleniumlibrary==6.8.0

# Parallel execution
robotframework-pabot==5.2.2

# Data-driven testing
robotframework-datadriver[XLS]==1.11.2

selenium==4.24.0
webdriver-manager==4.0.2

# Load config for env
python-dotenv==1.2.2